+++ custom_components/swetrack/const.py
@@ -8,6 +8,7 @@ PLATFORMS = ["device_tracker", "sensor", "binary_sensor"]
 CONF_API_KEY = "api_key"
 CONF_ACCOUNT_ID = "account_id"
 CONF_ACCOUNT_NAME = "account_name"
+CONF_ACCOUNT_LANGUAGE = "account_language"
 CONF_UPDATE_MODE = "update_mode"
 CONF_UPDATE_INTERVAL = "update_interval"
 CONF_ENABLED_DEVICES = "enabled_devices"